+++ example/lib/pages/location_stream_widget.dart
@@ -47,7 +47,7 @@ class LocationStreamState extends State<LocationStreamWidget> {
   @override
   Widget build(BuildContext context) {
     return FutureBuilder(
-        future: Geolocator.checkGeolocationStatus(),
+        future: Geolocator().checkGeolocationPermissionStatus(),
         builder:
             (BuildContext context, AsyncSnapshot<GeolocationStatus> snapshot) {
           if (!snapshot.hasData) {