+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.6.4]
 
 * Added support to supply a locale when using the `placemarkFromAddress` and `placemarkFromCoordinates` methods.
+* Depracated the static method `checkGeolocationStatus` in favor of the instance method `checkGeolocationPermissionStatus` (the static version will be removed in version 2.0 of the Geolocator plugin).
 
 ## [1.6.3]